x_langchain/data_analysis/pyproject.toml
[project]
name = "data-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.12",
    "langchain-modal>=0.0.2",
    "langchain-openai>=1.1.12",
    "matplotlib>=3.10.6",
    "modal>=1.4.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "slack-sdk>=3.41.0",
]
